ci: update release draft process in GitHub actions
Refactored the Create Release Draft step in our GitHub builds to use the nekofar/create-github-release action instead of manual scripting. This new action provides a more declarative and maintainable way to setup releases. It especially addresses the condition to mark the release as a pre-release in case the version contains '-rc', '-beta' or '-alpha'.